% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_fasta_example.R
\docType{data}
\name{human_fasta_example}
\alias{human_fasta_example}
\title{Imported Human Proteome (Fasta format)}
\format{A list of 20193 protein sequences and annotations}
\source{
\url{http://proteomecentral.proteomexchange.org/cgi/GetDataset?ID=PXD008079}
}
\usage{
human_fasta_example
}
\description{
Fasta formatted proteome for human proteins
}
\keyword{datasets}
