% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_evidence_example.R
\docType{data}
\name{malaria_evidence_example}
\alias{malaria_evidence_example}
\title{MaxQuant Evidence File}
\format{A data frame with 14687 rows and 103 columns}
\source{
\url{https://doi.org/10.1074/mcp.RA118.000907}
}
\usage{
malaria_evidence_example
}
\description{
A dataset containing post-translational modifications dervied from
Plasmodium falciparum-infectedhuman infected and non-infected humans. See
original publication for more details: Kawahara et al. MCP 2018
}
\keyword{datasets}
