% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_de.r
\name{phosphoDE}
\alias{phosphoDE}
\title{Perform Differential Phosphorylation Analysis (DPA)}
\usage{
phosphoDE(phospho_input = NULL, annotation_file = NULL,
  normalisation_method = "quantile", impute = FALSE, correct = FALSE,
  limma_analysis = TRUE, adjust_method = "BH", input_percent = 0.75,
  pairing = "unpaired", verbose = FALSE)
}
\arguments{
\item{phospho_input}{Data.frame of intensity values}

\item{annotation_file}{A data.frame describing the experimental design}

\item{normalisation_method}{Options are "none" or "quantile". Default =
"quantile"}

\item{impute}{Select from the following methods: "FALSE", "knn"}

\item{correct}{Performs batch correction using surrogate variable analysis
(SVA). Recommended to assess utility through PCA. Default = FALSE}

\item{limma_analysis}{Perform DPE using limma. Default = TRUE}

\item{adjust_method}{Method used for multiple comparison adjustment of
p-values. Options are: "holm", "hochberg", "hommel", "bonferroni", "BH",
"BY", "fdr" or "none". See ?p.adjust.methods for a details description and
references. Default = "BH"}

\item{input_percent}{Percentage of data that must be complete prior to
performing missing value imputation. Default = 0.75}

\item{pairing}{Default = "unpaired"}

\item{verbose}{Print progress to screen. Default = FALSE}
}
\value{
A List of 5 objects. 1) "results" are the limma results, 2)
"kinswingr_input" is a data.frame formatted for KinSwingR analysis, 3)
"corrected_data" is normalised + SVA corrected intensities 4)
"normalised_data" is normalised, before SVA and 5) "imputed_data" is imputed
data. Where an option was not selected, will result in NA.
}
\description{
Given a remapped evidence file from a MaxQuant output, this
function performs statistical analysis of phosphorylation data. It performs
normalisation, missing value imputation, batch correction utilising SVA and
statistical analysis of differential phosphorylation using limma. Currently
it only supports automated pairwise analysis between two groups. Will not run
limma if more than two groups detected. See limma_analysis parameter.
}
\examples{
## Use the outputs from the first example to demonstrate
## Return the annotated data with extracted peptides:
## load example data in phosphoProcessR
data(human_fasta_example)
data(malaria_evidence_example)
data(malaria_annotation_example)

## Take the first 100 for an example
evidence_file_example <- head(malaria_evidence_example, 100)

## convert evidence file into useful table for further analyses
evidence_tidy <- tidyEvidence(evidence_file = evidence_file_example,
                              annotation_file = malaria_annotation_example,
                              fasta_file = human_fasta_example,
                              window_size = 15,
                              min_prob = 0.7,
                              filter_site_method = "site")

## Run statistical analysis on tidy and remapped evidence file
DPA_out <- phosphoDE(phospho_input = evidence_tidy$intensity,
                     annotation_file = malaria_annotation_example,
                     correct = TRUE,
                     adjust_method = "none")

## View the corrected results
head(DPA_out$corrected_data)

}
