% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{plot_this}
\alias{plot_this}
\title{QC/diagnostic plotting}
\usage{
plot_this(data_in = NULL, annotation_file = NULL, legend = TRUE,
  label = TRUE, title = NA, pca = FALSE, hclust = FALSE,
  density = FALSE)
}
\arguments{
\item{data_in}{Either a numeric vector for plotting density distribution of
site probabilities of a data.frame that contains intensities.}

\item{annotation_file}{A data.frame describing the experimental design.}

\item{legend}{Include legend in plots? Legend is based on group data in
annotation_file Only used for PCA and hierarchical clustering. Options are:
TRUE, FALSE. Default = FALSE}

\item{label}{Include point labels in plots? Points are based on group data in
annotation_file. Options: TRUE, FALSE. Default = FALSE}

\item{title}{What to name the plot? Default = NULL}

\item{pca}{Perform unsupervised Principle Component Analysis (PCA) and plot
results. By default performs Singular Value Decomposition. Requires data_in
be a "data.frame" and utilises "label" meta-data from annotation_file for
colouring. Options:  TRUE, FALSE. Default = FALSE}

\item{hclust}{Performs unsupervised hierarchical clustering of samples.
Colours sample below plot according to group and numbered by inputs. Requires
data_in be a "data.frame" and utilises "label" meta-data from annotation_file
for colouring. Options:  TRUE, FALSE. Default = FALSE}

\item{density}{Plot density distributions of site probabilities. Input must
be a numerical vector. Options: TRUE, FALSE. Default = FALSE}
}
\value{
Returns pretty plots
}
\description{
Wrappers for a series of plots to be used as diagnostics in
in conjunction with phosphoProcessR analyses. Currently 3 plots are possible
using this function: 1) Principle Component Analyis (PCA), 2) Hierarchical
clustering, 3) density distribution of phosphosite probability. See
descriptions of each in the parameter options below and for format
specification. See vignette for more information and examples.
}
\examples{
## Load the example data set and attach

}
