% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_peptides.r
\name{tidyEvidence}
\alias{tidyEvidence}
\title{Remap and merge MaxQuant Evidence file into "tidy" data frame}
\usage{
tidyEvidence(evidence_file = NULL, annotation_file = NULL,
  fasta_file = NULL, window_size = 15, min_prob = 0,
  max_sites = "all", filter_site_method = "site",
  return_intensity = TRUE, return_mapping_table = TRUE,
  return_site_probability = TRUE, return_site_numbers = TRUE,
  verbose = FALSE)
}
\arguments{
\item{evidence_file}{MaxQuant evidence.txt file}

\item{annotation_file}{Table with annotations of the columns to extract}

\item{fasta_file}{Protein fasta file used for original mapping with MaxQuant}

\item{window_size}{The size of the AA window to be used for the remapping
around the phosphosite. E.g. XXXSXXX, would correspond to a window size of 7.
Default = 15}

\item{min_prob}{The minimum probability observed for any phosphosite in a
peptide. Recommended to use 0.7, but first assess with distribution as per
output table. See vignette for example. Default = 0}

\item{max_sites}{This will filter out peptides with numbers of phosphosites 
greater than this value. Options are numeric values (e.g., 1, 2, 3) or "all".
"all" keeps all peptides. Default="all"}

\item{filter_site_method}{Two methods for filtering peptides by site
probability are available. "site" will remove all phosphosites identified
with a probability lower that "min_prob". "peptide" will remove all peptides
that are predicted to contain a set combination of sites which do not conatin
a minimum probability of a site. Options are "site" or "peptide". Default =
"site"}

\item{return_intensity}{If TRUE will evaluate the MaxQuant evidence results
and return a table of intensities for further analysis. If FALSE, the
evidence file will not be "tidied" and the intensity object returned by 
tidyEvidence will be NA. Default = TRUE}

\item{return_mapping_table}{If TRUE a mapping table will be generated from 
the fasta file input and returned. If FALSE the mapping_table returned by 
tidyEvidence will be NA. Default = TRUE}

\item{return_site_probability}{If TRUE will evaluate the MaxQuant evidence 
file and return a vector of all the site probabilities identified in all 
phosphorylated peptides. If FALSE, the site_probability vector returned by
tidyEvidence will be NA. Default = TRUE.}

\item{return_site_numbers}{If TRUE will evaluate the MaxQuant evidence 
file and return a vector of the number of sites identified for each peptide. 
If FALSE, the site_probability vector returned by tidyEvidence will be NA. 
Default = TRUE.}

\item{verbose}{Print progress to screen. Default = FALSE}
}
\value{
List of 3 objects. 1) xx$intensity is a reformatted table of
intensities, now suitable for further statistical analyses, 2)
xx$mapping_table is table of remapped peptides and 3) xx$site_probability is a
vector of all the site probabilities identified.
}
\description{
Given a protein database, this function will remap and annotate
phosphorylated peptide sequence obtained from a MaxQuant evidence.txt file
ready for further statistical analysis
}
\examples{
## Return the annotated data with extracted peptides:
## load example data in phosphoProcessR
data(human_fasta_example)
data(malaria_evidence_example)
data(malaria_annotation_example)

## Take the first 100 for an example
evidence_file_example <- head(malaria_evidence_example, 100)

## convert evidence file into useful table for further analyses
evidence_tidy <- tidyEvidence(evidence_file = evidence_file_example,
                              annotation_file = malaria_annotation_example,
                              fasta_file = human_fasta_example,
                              window_size = 15,
                              min_prob = 0.7,
                              filter_site_method = "site")

## View results of table ready for statistical analysis
head(evidence_tidy$intensity, 5)

## Access site probabilities
head(evidence_tidy$site_probability)

}
